+++ apps/api/CHANGELOG.md
@@ -21,6 +21,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [2.3.1] - 2023-03-24
+
+### Fixed
+
+-   Sorting for images when filtering by modification times.
+
+## [2.3.0] - 2023-03-24
+
+### Added
+
+-   Filtering by dates (`created`, `modified`, `modifiedFile`) for images.
+
 ## [2.2.0] - 2023-03-21
 
 ### Changed