+++ apps/edit/CHANGELOG.md
@@ -21,6 +21,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [1.0.2] - 2023-03-24
+
+### Changed
+
+-   Updates for revisions to `SourceClient`.
+-   No longer updating `image` entity's `modified` timestamp when file is updated.
+
 ## [1.0.1] - 2023-03-10
 
 ### Fixed