+++ packages/source-client/CHANGELOG.md
@@ -19,4 +19,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [1.1.0] - 2023-03-24
+
+### Changed
+
+-   S3 Listers now return an object with a `Key` and `LastModified`, not just the key.
+
 ## [1.0.0] - 2023-02-12